requirements.txt
# Core AI / APIs
groq>=0.9.0
edge-tts>=6.1.9
googlesearch-python>=1.2.3

# Speech
SpeechRecognition>=3.10.0
pyaudio>=0.2.14

# Audio playback
pygame>=2.5.2

# GUI
PyQt5>=5.15.10

# Automation
pywhatkit>=5.4
AppOpener>=1.6
beautifulsoup4>=4.12.3
keyboard>=0.13.5
requests>=2.31.0

# Image generation
Pillow>=10.2.0

# Utilities
python-dotenv>=1.0.1
rich>=13.7.0
